+++ app/controllers/users_controller.rb
@@ -38,7 +38,7 @@ class UsersController < ApplicationController
   def update
     original_status = @user.status
     if @user.update(user_params)
-      UserMailer.status_activated(@user).deliver_later if Rails.env.production? and original_status != @user.status and @user.status == 'active'
+      UserMailer.status_activated(@user).deliver_later if EMAILS_ENABLED and original_status != @user.status and @user.status == 'active'
       redirect_to @user, notice: 'User was successfully updated.'
     else
       render action: 'edit'